Replace role-based alias config with mount based alias config (#91)
This moves the role-based alias configuration fields (`iam_alias`,
`gce_alias` from an individual role to the GCP configuration. This will
allow the user to specify the configuration for all roles rather than
per-role.
